+++ dnsimple/registrar_test.go
@@ -108,10 +108,10 @@ func TestRegistrarService_RegisterDomain(t *testing.T) {
 	}
 
 	registration := registrationResponse.Data
-	if want, got := 1, registration.ID; want != got {
+	if want, got := int64(1), registration.ID; want != got {
 		t.Fatalf("Registrar.RegisterDomain() returned ID expected to be `%v`, got `%v`", want, got)
 	}
-	if want, got := 999, registration.DomainID; want != got {
+	if want, got := int64(999), registration.DomainID; want != got {
 		t.Fatalf("Registrar.RegisterDomain() returned DomainID expected to be `%v`, got `%v`", want, got)
 	}
 }
@@ -165,10 +165,10 @@ func TestRegistrarService_TransferDomain(t *testing.T) {
 	}
 
 	transfer := transferResponse.Data
-	if want, got := 1, transfer.ID; want != got {
+	if want, got := int64(1), transfer.ID; want != got {
 		t.Fatalf("Registrar.TransferDomain() returned ID expected to be `%v`, got `%v`", want, got)
 	}
-	if want, got := 999, transfer.DomainID; want != got {
+	if want, got := int64(999), transfer.DomainID; want != got {
 		t.Fatalf("Registrar.TransferDomain() returned DomainID expected to be `%v`, got `%v`", want, got)
 	}
 }
@@ -237,9 +237,9 @@ func TestRegistrarService_GetDomainTransfer(t *testing.T) {
 	}
 	domainTransfer := domainTransferResponse.Data
 	wantSingle := &DomainTransfer{
-		ID:                361,
-		DomainID:          182245,
-		RegistrantID:      2715,
+		ID:                int64(361),
+		DomainID:          int64(182245),
+		RegistrantID:      int64(2715),
 		State:             "cancelled",
 		AutoRenew:         false,
 		WhoisPrivacy:      false,
@@ -272,9 +272,9 @@ func TestRegistrarService_CancelDomainTransfer(t *testing.T) {
 	}
 	domainTransfer := domainTransferResponse.Data
 	wantSingle := &DomainTransfer{
-		ID:                361,
-		DomainID:          182245,
-		RegistrantID:      2715,
+		ID:                int64(361),
+		DomainID:          int64(182245),
+		RegistrantID:      int64(2715),
 		State:             "transferring",
 		AutoRenew:         false,
 		WhoisPrivacy:      false,
@@ -310,10 +310,10 @@ func TestRegistrarService_RenewDomain(t *testing.T) {
 	}
 
 	renewal := renewalResponse.Data
-	if want, got := 1, renewal.ID; want != got {
+	if want, got := int64(1), renewal.ID; want != got {
 		t.Fatalf("Registrar.RenewDomain() returned ID expected to be `%v`, got `%v`", want, got)
 	}
-	if want, got := 999, renewal.DomainID; want != got {
+	if want, got := int64(999), renewal.DomainID; want != got {
 		t.Fatalf("Registrar.RenewDomain() returned DomainID expected to be `%v`, got `%v`", want, got)
 	}
 }
